Death Controller, persister yml and migration script, ack no ,uuid, mdms util ,repositories and query builder

+++ municipal-services/death-services/src/main/java/org/ksmart/death/deathapplication/web/models/DeathBasicInfo.java
@@ -344,7 +344,7 @@ public class DeathBasicInfo {
 
     @Schema(type = "Integer", description= "ack no id" )
     @JsonProperty("ackNoID")
-    private Integer  ackNoID ;
+    private Long  ackNoID ;
 
     @Schema(type = "Long",description = "Death Application Date  ")
     @JsonProperty("ApplicationDate")
